requirements.txt
Django==1.5.1
colorama==0.2.5
cx-Oracle==5.1.2
django-tastypie==0.11.0
ipython==0.13.2
jira-python==0.13
oauthlib==0.4.1
psycopg2==2.5
pychecker==0.8.19
pyreadline==2.0
python-dateutil==2.2
python-magic==0.4.3
python-mimeparse==0.1.4
requests==1.2.2
requests-oauthlib==0.3.2
selenium==2.30.0
six==1.5.2
tlslite==0.4.1
virtualenv==1.9.1
